// interfaces/futurama-list.ts
// Generated by https://quicktype.io

export interface FuturamaList {
  name:        Name;
  images:      Images;
  gender:      Gender;
  species:     string;
  homePlanet?: string;
  occupation:  string;
  sayings:     string[];
  id:          number;
  age:         string;
}

export enum Gender {
  Female = "Female",
  Male = "Male",
}

export interface Images {
  "head-shot": string;
  main:        string;
}

export interface Name {
  first:  string;
  middle: string;
  last:   string;
}

export interface SmallFuturama {
  name: string,
  age: string,
  image: string,
  id: number
}